Snapshot of work performed towards late safepoint placement since initial release.  This change is a diff of all changes (including previous released ones) against upstream LLVM on July 1st.

+++ tools/bugpoint/OptimizerDriver.cpp
@@ -39,6 +39,11 @@ using namespace llvm;
 #define DEBUG_TYPE "bugpoint"
 
 namespace llvm {
+  cl::opt<bool>
+  KeepTemps("bugpoint-keep-temps",
+                    cl::desc("Keep all the temporary input and output files used by bugpoint"),
+                    cl::init(false));
+
   extern cl::opt<std::string> OutputPrefix;
 }
 
@@ -220,13 +225,16 @@ bool BugDriver::runPasses(Module *Program,
                                    Timeout, MemoryLimit, &ErrMsg);
 
   // If we are supposed to delete the bitcode file or if the passes crashed,
-  // remove it now.  This may fail if the file was never created, but that's ok.
-  if (DeleteOutput || result != 0)
-    sys::fs::remove(OutputFilename);
-
-  // Remove the temporary input file as well
-  sys::fs::remove(InputFilename.c_str());
-
+  // remove it now.  This may fail if the file was never created, but that's
+  // ok.
+  if( !KeepTemps ) {
+    if (DeleteOutput || result != 0)
+      sys::fs::remove(OutputFilename);
+    
+    // Remove the temporary input file as well
+    sys::fs::remove(InputFilename.c_str());
+  }
+  
   if (!Quiet) {
     if (result == 0)
       outs() << "Success!\n";
